feat!: remove all sync APIs, rename async classes
BREAKING CHANGE: All synchronous API classes have been removed.
The async classes are now the only implementation and have been
renamed to drop the Async prefix:
- AsyncNextcloud -> Nextcloud
- AsyncNextcloudApp -> NextcloudApp
- AsyncFilesAPI -> FilesAPI
- AsyncTalkBot -> TalkBot
- anc_app -> nc_app
- atalk_bot_msg -> talk_bot_msg

Backward compatibility aliases are provided and will be removed
in v1.0.0.

Removed ~2,950 lines of duplicated sync code.
