+++ nc_py_api/ex_app/providers/providers.py
@@ -1,7 +1,7 @@
 """Nextcloud API for AI Providers."""
 
-from ..._session import AsyncNcSessionApp, NcSessionApp
-from .task_processing import _AsyncTaskProcessingProviderAPI, _TaskProcessingProviderAPI
+from ..._session import AsyncNcSessionApp
+from .task_processing import _TaskProcessingProviderAPI
 
 
 class ProvidersApi:
@@ -10,15 +10,5 @@ class ProvidersApi:
     task_processing: _TaskProcessingProviderAPI
     """TaskProcessing Provider API."""
 
-    def __init__(self, session: NcSessionApp):
-        self.task_processing = _TaskProcessingProviderAPI(session)
-
-
-class AsyncProvidersApi:
-    """Class that encapsulates all AI Providers functionality."""
-
-    task_processing: _AsyncTaskProcessingProviderAPI
-    """TaskProcessing Provider API."""
-
     def __init__(self, session: AsyncNcSessionApp):
-        self.task_processing = _AsyncTaskProcessingProviderAPI(session)
+        self.task_processing = _TaskProcessingProviderAPI(session)
