+++ nc_py_api/ex_app/providers/task_processing.py
@@ -10,7 +10,7 @@ from pydantic.dataclasses import dataclass
 
 from ..._exceptions import NextcloudException, NextcloudExceptionNotFound
 from ..._misc import require_capabilities
-from ..._session import AsyncNcSessionApp, NcSessionApp
+from ..._session import AsyncNcSessionApp
 
 _EP_SUFFIX: str = "ai_provider/task_processing"
 
@@ -104,101 +104,6 @@ class TaskProcessingProvider:
 
 
 class _TaskProcessingProviderAPI:
-    """API for TaskProcessing providers, available as **nc.providers.task_processing.<method>**."""
-
-    def __init__(self, session: NcSessionApp):
-        self._session = session
-
-    def register(
-        self,
-        provider: TaskProcessingProvider,
-        custom_task_type: TaskType | None = None,
-    ) -> None:
-        """Registers or edit the TaskProcessing provider."""
-        require_capabilities("app_api", self._session.capabilities)
-        params = {
-            "provider": RootModel(provider).model_dump(),
-            **({"customTaskType": RootModel(custom_task_type).model_dump()} if custom_task_type else {}),
-        }
-        self._session.ocs("POST", f"{self._session.ae_url}/{_EP_SUFFIX}", json=params)
-
-    def unregister(self, name: str, not_fail=True) -> None:
-        """Removes TaskProcessing provider."""
-        require_capabilities("app_api", self._session.capabilities)
-        try:
-            self._session.ocs("DELETE", f"{self._session.ae_url}/{_EP_SUFFIX}", params={"name": name})
-        except NextcloudExceptionNotFound as e:
-            if not not_fail:
-                raise e from None
-
-    def next_task(self, provider_ids: list[str], task_types: list[str]) -> dict[str, typing.Any]:
-        """Get the next task processing task from Nextcloud."""
-        with contextlib.suppress(NextcloudException):
-            if r := self._session.ocs(
-                "GET",
-                "/ocs/v2.php/taskprocessing/tasks_provider/next",
-                json={"providerIds": provider_ids, "taskTypeIds": task_types},
-            ):
-                return r
-        return {}
-
-    def next_task_batch(
-        self, provider_ids: list[str], task_types: list[str], number_of_tasks: int
-    ) -> dict[str, typing.Any]:
-        """Get the next n task processing tasks from Nextcloud.
-
-        Available starting with Nextcloud 33
-        Returns: {tasks: [{task: Task, provider: string}], has_more: bool}
-        """
-        with contextlib.suppress(NextcloudException):
-            if r := self._session.ocs(
-                "GET",
-                "/ocs/v2.php/taskprocessing/tasks_provider/next_batch",
-                json={"providerIds": provider_ids, "taskTypeIds": task_types, "numberOfTasks": number_of_tasks},
-            ):
-                return r
-        return {"tasks": [], "has_more": False}
-
-    def set_progress(self, task_id: int, progress: float) -> dict[str, typing.Any]:
-        """Report new progress value of the task to Nextcloud. Progress should be in range from 0.0 to 100.0."""
-        with contextlib.suppress(NextcloudException):
-            if r := self._session.ocs(
-                "POST",
-                f"/ocs/v2.php/taskprocessing/tasks_provider/{task_id}/progress",
-                json={"taskId": task_id, "progress": progress / 100.0},
-            ):
-                return r
-        return {}
-
-    def upload_result_file(self, task_id: int, file: bytes | str | typing.Any) -> int:
-        """Uploads file and returns fileID that should be used in the ``report_result`` function.
-
-        .. note:: ``file`` can be any file-like object.
-        """
-        return self._session.ocs(
-            "POST",
-            f"/ocs/v2.php/taskprocessing/tasks_provider/{task_id}/file",
-            files={"file": file},
-        )["fileId"]
-
-    def report_result(
-        self,
-        task_id: int,
-        output: dict[str, typing.Any] | None = None,
-        error_message: str | None = None,
-    ) -> dict[str, typing.Any]:
-        """Report result of the task processing to Nextcloud."""
-        with contextlib.suppress(NextcloudException):
-            if r := self._session.ocs(
-                "POST",
-                f"/ocs/v2.php/taskprocessing/tasks_provider/{task_id}/result",
-                json={"taskId": task_id, "output": output, "errorMessage": error_message},
-            ):
-                return r
-        return {}
-
-
-class _AsyncTaskProcessingProviderAPI:
     """Async API for TaskProcessing providers."""
 
     def __init__(self, session: AsyncNcSessionApp):
