+++ nc_py_api/ex_app/ui/resources.py
@@ -4,7 +4,7 @@ import dataclasses
 
 from ..._exceptions import NextcloudExceptionNotFound
 from ..._misc import require_capabilities
-from ..._session import AsyncNcSessionApp, NcSessionApp
+from ..._session import AsyncNcSessionApp
 
 
 @dataclasses.dataclass
@@ -77,130 +77,6 @@ class UiStyle(UiBase):
 
 
 class _UiResources:
-    """API for adding scripts, styles, initial-states to the pages, avalaible as **nc.ui.resources.<method>**."""
-
-    _ep_suffix_init_state: str = "ui/initial-state"
-    _ep_suffix_js: str = "ui/script"
-    _ep_suffix_css: str = "ui/style"
-
-    def __init__(self, session: NcSessionApp):
-        self._session = session
-
-    def set_initial_state(self, ui_type: str, name: str, key: str, value: dict | list) -> None:
-        """Add or update initial state for the page(template)."""
-        require_capabilities("app_api", self._session.capabilities)
-        params = {
-            "type": ui_type,
-            "name": name,
-            "key": key,
-            "value": value,
-        }
-        self._session.ocs("POST", f"{self._session.ae_url}/{self._ep_suffix_init_state}", json=params)
-
-    def delete_initial_state(self, ui_type: str, name: str, key: str, not_fail=True) -> None:
-        """Removes initial state for the page(template) by object name."""
-        require_capabilities("app_api", self._session.capabilities)
-        try:
-            self._session.ocs(
-                "DELETE",
-                f"{self._session.ae_url}/{self._ep_suffix_init_state}",
-                params={"type": ui_type, "name": name, "key": key},
-            )
-        except NextcloudExceptionNotFound as e:
-            if not not_fail:
-                raise e from None
-
-    def get_initial_state(self, ui_type: str, name: str, key: str) -> UiInitState | None:
-        """Get information about initial state for the page(template) by object name."""
-        require_capabilities("app_api", self._session.capabilities)
-        try:
-            return UiInitState(
-                self._session.ocs(
-                    "GET",
-                    f"{self._session.ae_url}/{self._ep_suffix_init_state}",
-                    params={"type": ui_type, "name": name, "key": key},
-                )
-            )
-        except NextcloudExceptionNotFound:
-            return None
-
-    def set_script(self, ui_type: str, name: str, path: str, after_app_id: str = "") -> None:
-        """Add or update script for the page(template)."""
-        require_capabilities("app_api", self._session.capabilities)
-        params = {
-            "type": ui_type,
-            "name": name,
-            "path": path,
-            "afterAppId": after_app_id,
-        }
-        self._session.ocs("POST", f"{self._session.ae_url}/{self._ep_suffix_js}", json=params)
-
-    def delete_script(self, ui_type: str, name: str, path: str, not_fail=True) -> None:
-        """Removes script for the page(template) by object name."""
-        require_capabilities("app_api", self._session.capabilities)
-        try:
-            self._session.ocs(
-                "DELETE",
-                f"{self._session.ae_url}/{self._ep_suffix_js}",
-                params={"type": ui_type, "name": name, "path": path},
-            )
-        except NextcloudExceptionNotFound as e:
-            if not not_fail:
-                raise e from None
-
-    def get_script(self, ui_type: str, name: str, path: str) -> UiScript | None:
-        """Get information about script for the page(template) by object name."""
-        require_capabilities("app_api", self._session.capabilities)
-        try:
-            return UiScript(
-                self._session.ocs(
-                    "GET",
-                    f"{self._session.ae_url}/{self._ep_suffix_js}",
-                    params={"type": ui_type, "name": name, "path": path},
-                )
-            )
-        except NextcloudExceptionNotFound:
-            return None
-
-    def set_style(self, ui_type: str, name: str, path: str) -> None:
-        """Add or update style(css) for the page(template)."""
-        require_capabilities("app_api", self._session.capabilities)
-        params = {
-            "type": ui_type,
-            "name": name,
-            "path": path,
-        }
-        self._session.ocs("POST", f"{self._session.ae_url}/{self._ep_suffix_css}", json=params)
-
-    def delete_style(self, ui_type: str, name: str, path: str, not_fail=True) -> None:
-        """Removes style(css) for the page(template) by object name."""
-        require_capabilities("app_api", self._session.capabilities)
-        try:
-            self._session.ocs(
-                "DELETE",
-                f"{self._session.ae_url}/{self._ep_suffix_css}",
-                params={"type": ui_type, "name": name, "path": path},
-            )
-        except NextcloudExceptionNotFound as e:
-            if not not_fail:
-                raise e from None
-
-    def get_style(self, ui_type: str, name: str, path: str) -> UiStyle | None:
-        """Get information about style(css) for the page(template) by object name."""
-        require_capabilities("app_api", self._session.capabilities)
-        try:
-            return UiStyle(
-                self._session.ocs(
-                    "GET",
-                    f"{self._session.ae_url}/{self._ep_suffix_css}",
-                    params={"type": ui_type, "name": name, "path": path},
-                )
-            )
-        except NextcloudExceptionNotFound:
-            return None
-
-
-class _AsyncUiResources:
     """Async API for adding scripts, styles, initial-states to the TopMenu pages."""
 
     _ep_suffix_init_state: str = "ui/initial-state"
