+++ nc_py_api/ex_app/ui/top_menu.py
@@ -4,7 +4,7 @@ import dataclasses
 
 from ..._exceptions import NextcloudExceptionNotFound
 from ..._misc import require_capabilities
-from ..._session import AsyncNcSessionApp, NcSessionApp
+from ..._session import AsyncNcSessionApp
 
 
 @dataclasses.dataclass
@@ -44,51 +44,6 @@ class UiTopMenuEntry:
 
 
 class _UiTopMenuAPI:
-    """API for the top menu app nav bar in Nextcloud, avalaible as **nc.ui.top_menu.<method>**."""
-
-    _ep_suffix: str = "ui/top-menu"
-
-    def __init__(self, session: NcSessionApp):
-        self._session = session
-
-    def register(self, name: str, display_name: str, icon: str = "", admin_required=False) -> None:
-        """Registers or edit the App entry in Top Meny.
-
-        :param name: Unique name for the menu entry.
-        :param display_name: Display name of the menu entry.
-        :param icon: Optional, url relative to the ExApp, like: "img/icon.svg"
-        :param admin_required: Boolean value indicating should be Entry visible to all or only to admins.
-        """
-        require_capabilities("app_api", self._session.capabilities)
-        params = {
-            "name": name,
-            "displayName": display_name,
-            "icon": icon,
-            "adminRequired": int(admin_required),
-        }
-        self._session.ocs("POST", f"{self._session.ae_url}/{self._ep_suffix}", json=params)
-
-    def unregister(self, name: str, not_fail=True) -> None:
-        """Removes App entry in Top Menu."""
-        require_capabilities("app_api", self._session.capabilities)
-        try:
-            self._session.ocs("DELETE", f"{self._session.ae_url}/{self._ep_suffix}", params={"name": name})
-        except NextcloudExceptionNotFound as e:
-            if not not_fail:
-                raise e from None
-
-    def get_entry(self, name: str) -> UiTopMenuEntry | None:
-        """Get information of the top meny entry for current app."""
-        require_capabilities("app_api", self._session.capabilities)
-        try:
-            return UiTopMenuEntry(
-                self._session.ocs("GET", f"{self._session.ae_url}/{self._ep_suffix}", params={"name": name})
-            )
-        except NextcloudExceptionNotFound:
-            return None
-
-
-class _AsyncUiTopMenuAPI:
     """Async API for the top menu app nav bar in Nextcloud."""
 
     _ep_suffix: str = "ui/top-menu"
