+++ nc_py_api/ex_app/ui/ui.py
@@ -1,10 +1,10 @@
 """Nextcloud API for User Interface."""
 
-from ..._session import AsyncNcSessionApp, NcSessionApp
-from .files_actions import _AsyncUiFilesActionsAPI, _UiFilesActionsAPI
-from .resources import _AsyncUiResources, _UiResources
-from .settings import _AsyncDeclarativeSettingsAPI, _DeclarativeSettingsAPI
-from .top_menu import _AsyncUiTopMenuAPI, _UiTopMenuAPI
+from ..._session import AsyncNcSessionApp
+from .files_actions import _UiFilesActionsAPI
+from .resources import _UiResources
+from .settings import _DeclarativeSettingsAPI
+from .top_menu import _UiTopMenuAPI
 
 
 class UiApi:
@@ -19,27 +19,8 @@ class UiApi:
     settings: _DeclarativeSettingsAPI
     """API for ExApp settings UI"""
 
-    def __init__(self, session: NcSessionApp):
+    def __init__(self, session: AsyncNcSessionApp):
         self.files_dropdown_menu = _UiFilesActionsAPI(session)
         self.top_menu = _UiTopMenuAPI(session)
         self.resources = _UiResources(session)
         self.settings = _DeclarativeSettingsAPI(session)
-
-
-class AsyncUiApi:
-    """Class that encapsulates all UI functionality(async)."""
-
-    files_dropdown_menu: _AsyncUiFilesActionsAPI
-    """File dropdown menu API."""
-    top_menu: _AsyncUiTopMenuAPI
-    """Top App menu API."""
-    resources: _AsyncUiResources
-    """Page(Template) resources API."""
-    settings: _AsyncDeclarativeSettingsAPI
-    """API for ExApp settings UI"""
-
-    def __init__(self, session: AsyncNcSessionApp):
-        self.files_dropdown_menu = _AsyncUiFilesActionsAPI(session)
-        self.top_menu = _AsyncUiTopMenuAPI(session)
-        self.resources = _AsyncUiResources(session)
-        self.settings = _AsyncDeclarativeSettingsAPI(session)
