+++ nc_py_api/notifications.py
@@ -12,8 +12,6 @@ from ._misc import (
 from ._session import (
     AsyncNcSessionApp,
     AsyncNcSessionBasic,
-    NcSessionApp,
-    NcSessionBasic,
 )
 
 
@@ -82,73 +80,6 @@ class Notification:
 
 
 class _NotificationsAPI:
-    """Class providing an API for managing user notifications on the Nextcloud server."""
-
-    _ep_base: str = "/ocs/v2.php/apps/notifications/api/v2/notifications"
-
-    def __init__(self, session: NcSessionBasic):
-        self._session = session
-
-    @property
-    def available(self) -> bool:
-        """Returns True if the Nextcloud instance supports this feature, False otherwise."""
-        return not check_capabilities("notifications", self._session.capabilities)
-
-    def create(
-        self,
-        subject: str,
-        message: str = "",
-        subject_params: dict | None = None,
-        message_params: dict | None = None,
-        link: str = "",
-    ) -> str:
-        """Create a Notification for the current user and returns it's ObjectID.
-
-        .. note:: Does not work in Nextcloud client mode, only for NextcloudApp mode.
-        """
-        params = _create(subject, message, subject_params, message_params, link)
-        if not isinstance(self._session, NcSessionApp):
-            raise NotImplementedError("Sending notifications is only supported for `App` mode.")
-        require_capabilities(["app_api", "notifications"], self._session.capabilities)
-        return self._session.ocs("POST", f"{self._session.ae_url}/notification", json=params)["object_id"]
-
-    def get_all(self) -> list[Notification]:
-        """Gets all notifications for a current user."""
-        require_capabilities("notifications", self._session.capabilities)
-        return [Notification(i) for i in self._session.ocs("GET", self._ep_base)]
-
-    def get_one(self, notification_id: int) -> Notification:
-        """Gets a single notification for a current user."""
-        require_capabilities("notifications", self._session.capabilities)
-        return Notification(self._session.ocs("GET", f"{self._ep_base}/{notification_id}"))
-
-    def by_object_id(self, object_id: str) -> Notification | None:
-        """Returns Notification if any by its object ID.
-
-        .. note:: this method is a temporary workaround until `create` can return `notification_id`.
-        """
-        for i in self.get_all():
-            if i.object_id == object_id:
-                return i
-        return None
-
-    def delete(self, notification_id: int) -> None:
-        """Deletes a notification for the current user."""
-        require_capabilities("notifications", self._session.capabilities)
-        self._session.ocs("DELETE", f"{self._ep_base}/{notification_id}")
-
-    def delete_all(self) -> None:
-        """Deletes all notifications for the current user."""
-        require_capabilities("notifications", self._session.capabilities)
-        self._session.ocs("DELETE", self._ep_base)
-
-    def exists(self, notification_ids: list[int]) -> list[int]:
-        """Checks the existence of notifications for the current user."""
-        require_capabilities("notifications", self._session.capabilities)
-        return self._session.ocs("POST", f"{self._ep_base}/exists", json={"ids": notification_ids})
-
-
-class _AsyncNotificationsAPI:
     """Class provides async API for managing user notifications on the Nextcloud server."""
 
     _ep_base: str = "/ocs/v2.php/apps/notifications/api/v2/notifications"
