+++ nc_py_api/users_groups.py
@@ -3,7 +3,7 @@
 import dataclasses
 
 from ._misc import kwargs_to_params
-from ._session import AsyncNcSessionBasic, NcSessionBasic
+from ._session import AsyncNcSessionBasic
 
 
 @dataclasses.dataclass
@@ -48,57 +48,6 @@ class GroupDetails:
 
 
 class _UsersGroupsAPI:
-    """Class providing an API for managing user groups on the Nextcloud server.
-
-    .. note:: In NextcloudApp mode, only ``get_list`` and ``get_details`` methods are available.
-    """
-
-    _ep_base: str = "/ocs/v1.php/cloud/groups"
-
-    def __init__(self, session: NcSessionBasic):
-        self._session = session
-
-    def get_list(self, mask: str | None = None, limit: int | None = None, offset: int | None = None) -> list[str]:
-        """Returns a list of user groups IDs."""
-        data = kwargs_to_params(["search", "limit", "offset"], search=mask, limit=limit, offset=offset)
-        response_data = self._session.ocs("GET", self._ep_base, params=data)
-        return response_data["groups"] if response_data else []
-
-    def get_details(
-        self, mask: str | None = None, limit: int | None = None, offset: int | None = None
-    ) -> list[GroupDetails]:
-        """Returns a list of user groups with detailed information."""
-        data = kwargs_to_params(["search", "limit", "offset"], search=mask, limit=limit, offset=offset)
-        response_data = self._session.ocs("GET", f"{self._ep_base}/details", params=data)
-        return [GroupDetails(i) for i in response_data["groups"]] if response_data else []
-
-    def create(self, group_id: str, display_name: str | None = None) -> None:
-        """Creates the users group."""
-        params = {"groupid": group_id}
-        if display_name is not None:
-            params["displayname"] = display_name
-        self._session.ocs("POST", f"{self._ep_base}", params=params)
-
-    def edit(self, group_id: str, display_name: str) -> None:
-        """Edits users group information."""
-        params = {"key": "displayname", "value": display_name}
-        self._session.ocs("PUT", f"{self._ep_base}/{group_id}", params=params)
-
-    def delete(self, group_id: str) -> None:
-        """Removes the users group."""
-        self._session.ocs("DELETE", f"{self._ep_base}/{group_id}")
-
-    def get_members(self, group_id: str) -> list[str]:
-        """Returns a list of group users."""
-        response_data = self._session.ocs("GET", f"{self._ep_base}/{group_id}")
-        return response_data["users"] if response_data else {}
-
-    def get_subadmins(self, group_id: str) -> list[str]:
-        """Returns list of users who is subadmins of the group."""
-        return self._session.ocs("GET", f"{self._ep_base}/{group_id}/subadmins")
-
-
-class _AsyncUsersGroupsAPI:
     """Class provides an async API for managing user groups on the Nextcloud server.
 
     .. note:: In NextcloudApp mode, only ``get_list`` and ``get_details`` methods are available.
