+++ nc_py_api/webhooks.py
@@ -4,7 +4,7 @@ import dataclasses
 
 from ._exceptions import NextcloudExceptionNotFound
 from ._misc import clear_from_params_empty  # , require_capabilities
-from ._session import AppConfig, AsyncNcSessionBasic, NcSessionBasic
+from ._session import AppConfig, AsyncNcSessionBasic
 
 
 @dataclasses.dataclass
@@ -74,85 +74,6 @@ class WebhookInfo:
 
 
 class _WebhooksAPI:
-    """The class provides the application management API on the Nextcloud server."""
-
-    _ep_base: str = "/ocs/v1.php/apps/webhook_listeners/api/v1/webhooks"
-
-    def __init__(self, session: NcSessionBasic):
-        self._session = session
-
-    def get_list(self, uri_filter: str = "") -> list[WebhookInfo]:
-        params = {"uri": uri_filter} if uri_filter else {}
-        return [WebhookInfo(i) for i in self._session.ocs("GET", f"{self._ep_base}", params=params)]
-
-    def get_entry(self, webhook_id: int) -> WebhookInfo | None:
-        try:
-            return WebhookInfo(self._session.ocs("GET", f"{self._ep_base}/{webhook_id}"))
-        except NextcloudExceptionNotFound:
-            return None
-
-    def register(
-        self,
-        http_method: str,
-        uri: str,
-        event: str,
-        event_filter: dict | None = None,
-        user_id_filter: str = "",
-        headers: dict | None = None,
-        auth_method: str = "none",
-        auth_data: dict | None = None,
-    ):
-        params = {
-            "httpMethod": http_method,
-            "uri": uri,
-            "event": event,
-            "eventFilter": event_filter,
-            "userIdFilter": user_id_filter,
-            "headers": headers,
-            "authMethod": auth_method,
-            "authData": auth_data,
-        }
-        clear_from_params_empty(["eventFilter", "userIdFilter", "headers", "authMethod", "authData"], params)
-        return WebhookInfo(self._session.ocs("POST", f"{self._ep_base}", json=params))
-
-    def update(
-        self,
-        webhook_id: int,
-        http_method: str,
-        uri: str,
-        event: str,
-        event_filter: dict | None = None,
-        user_id_filter: str = "",
-        headers: dict | None = None,
-        auth_method: str = "none",
-        auth_data: dict | None = None,
-    ):
-        params = {
-            "id": webhook_id,
-            "httpMethod": http_method,
-            "uri": uri,
-            "event": event,
-            "eventFilter": event_filter,
-            "userIdFilter": user_id_filter,
-            "headers": headers,
-            "authMethod": auth_method,
-            "authData": auth_data,
-        }
-        clear_from_params_empty(["eventFilter", "userIdFilter", "headers", "authMethod", "authData"], params)
-        return WebhookInfo(self._session.ocs("POST", f"{self._ep_base}/{webhook_id}", json=params))
-
-    def unregister(self, webhook_id: int) -> bool:
-        return self._session.ocs("DELETE", f"{self._ep_base}/{webhook_id}")
-
-    def unregister_all(self, appid: str = "") -> int:
-        if not appid and isinstance(self._session.cfg, AppConfig):
-            appid = self._session.cfg.app_name
-        else:
-            raise ValueError("The `appid` parameter cannot be empty for non-ExApp use.")
-        return self._session.ocs("DELETE", f"{self._ep_base}/byappid/{appid}")
-
-
-class _AsyncWebhooksAPI:
     """The class provides the async application management API on the Nextcloud server."""
 
     _ep_base: str = "/ocs/v1.php/apps/webhook_listeners/api/v1/webhooks"
